requirements.txt
customtkinter
pystray
Pillow
requests
